requirements.txt
httpx==0.27.*
jinja2==3.1.*
python-dotenv==1.2.*
